Model cast count as integer

+++ lib/Model.php
@@ -1541,7 +1541,7 @@ class Model
 		$table = static::table();
 		$sql = $table->options_to_sql($options);
 		$values = $sql->get_where_values();
-		return static::connection()->query_and_fetch_one($sql->to_s(),$values);
+		return (int) static::connection()->query_and_fetch_one($sql->to_s(),$values);
 	}
 
 	/**
